+++ src/app/layout.tsx
@@ -1,8 +1,8 @@
-import "@/styles/global.scss";
+import "@/app/styles/global.scss";
 import type { Metadata } from "next";
 import { Poppins } from "next/font/google";
 import { ReactNode } from "react";
-import Providers from "../components/providers/provider";
+import Providers from "@/app/components/providers/Providers";
 
 const poppins = Poppins({
   weight: ["400", "700"],
